requirements.txt
amqp==2.5.2
asgiref==3.2.7
billiard==3.6.3.0
celery==4.4.2
certifi==2020.4.5.1
chardet==3.0.4
Django==3.0.6
django-crispy-forms==1.9.1
django-mathfilters==1.0.0
djangorestframework==3.11.0
gunicorn==20.0.4
idna==2.9
kombu==4.6.8
Pillow==7.1.2
pyTelegramBotAPI==3.7.1
pytz==2020.1
redis==3.5.2
requests==2.23.0
six==1.14.0
sqlparse==0.3.1
urllib3==1.25.9
vine==1.3.0
